sw_services: xilsecure:Fix for Misra C rule 14.3
Misra C violation rule 14.3:has a bigger upper bound than the type u8 of loop counter Index, the loop condition may always be true.

Soln:Fixed by adding typecast to u32


Acked-for-series: Durga Challa <durga.challa@amd.com>

+++ lib/sw_services/xilsecure/src/versal_gen/client/core/mailbox/xsecure_mailbox.c
@@ -65,7 +65,7 @@
 int XSecure_InputSlrIndex(XSecure_ClientInstance *InstancePtr, u32 SlrIndex)
 {
 	int Status = XST_FAILURE;
-	if(SlrIndex >= XSECURE_SLR_INDEX_0 && SlrIndex <= XSECURE_SLR_INDEX_3){
+	if((SlrIndex >= (u32)XSECURE_SLR_INDEX_0) && (SlrIndex <= (u32)XSECURE_SLR_INDEX_3)){
 		InstancePtr->SlrIndex = SlrIndex;
 	    Status =  XST_SUCCESS;
 	}
